requirements.txt
Pillow~=10.1.0
PyMySQL~=1.0.3
boto3==1.24.25
botocore==1.27.25
cryptography~=37.0.4
pytz~=2021.3
s3transfer~=0.6.2
tld~=0.13
validators~=0.20.0
